doc: add documentation for aggregate module
