+++ eventually/src/lib.rs
@@ -1,11 +1,12 @@
-#![allow(warnings)]
+#![allow(type_alias_bounds)]
 
 pub mod aggregate;
 pub mod command;
 pub mod store;
+pub mod versioned;
 
 pub use {
-    aggregate::{util::AggregateExt, Aggregate},
+    aggregate::{Aggregate, AggregateExt},
     command::Handler as CommandHandler,
     store::{ReadStore, WriteStore},
 };
